client: external-match-client: Add `GetFees` method to fetch match fee
